Add support for DigitalNZ v3 API
This adds a new module, `Net::DigitalNZ::V3`, that supports the current v3
DigitalNZ API.

The module can either be used on its own, or via the original module by passing
`version => 3` to the constructor.

Currently this only supports the search (which doubles as a harvesting API) and
metadata (fetching record details) endpoints. The "sets" and "more like this"
endpoints may be added if they're needed.